[FSDP2][ez] Combined communication test files
[ghstack-poisoned]
